Fixed a few win32 keyboard/mouse problems.  First, the mouse hook now captures non-client area mouse messages.  Previously, these were ignored (because i forgot about them) and they caused all kinds of problems because they weren't forwarded.  For example, clicking on a window border would cause the window to start resizing when the mouse came back to the server screen.  Moving inside a title bar meant that the mouse wouldn't move on the client screen.
Second, because non-client messages are now handled, the full
screen transparent window is no longer necessary to capture
input so it's never displayed.  (The window is still necessary
for clipboard ownership so it's still created.)  No transparent
window means no screen flashing.  It also means we don't have to
become the foreground and active window.  This plays better with
apps that minimize or restore when they're no longer the
foreground application/active window.

Third, fixed the low level keyboard hook to forward toggle key
updates, which it was neglecting to do.

Finally, keyboard and mouse input is always forwarded from the hook
to the primary screen handler which then shadows the current key
and mouse button state.  If we're using low level hooks then this
isn't really necessary and GetKeyState() always returns the right
info but without low level hooks it means we can just use the
shadow state.  It also means we don't have to show our window in
order to get the system's key state table up to date, fixing the
screen flash when checking for the scroll lock state.