Attempt to fix issue with \ and | on win8 client
This has not been tested as I do not have a windows build environment but I think it might work.  Pressing the backslash or bar key results in ctrl and shift keypresses on the client and stops it displaying the correct character in certain apps, e.g. cygwin, virtual box, rdp.  Please could you make a win8 binary available for me to test out the changes?