Fix about dialog appearance on linux breeze dark theme.
If breeze dark desktop theme is used on linux the about dialog
icon "Synergy" text is not legible. The height of the dialog is
also too short cutting off the last line of about text.
Remove "Synergy" from about.png and put label with 20pt text in
the theme foreground color instead.
Set height of about dialog on linux to 350 to make last line visible.