Add CMake option to use system PugiXML
Currently, synergy-core builds against a version of PugiXML included in a
submodule of the repository (and not included in the release tarballs at all,
so they actually can't be built). However, users who use a package management
system such as `brew` may prefer to centrally manage dependencies such as
PugiXML so that they can be tracked, versioned, and updated through the
package management system.

This commit adds a new CMake option, USE_SYSTEM_PUGIXML. When the option is
set to ON, synergy-core will build against the system's locally installed
version of PugiXML (found using `find_package`). The default is OFF, so there
is no change of behavior except for users who choose to set the option ON.