fix(gui): Restart Qt app instance without creating a new process
fixes: #7896

Instead of launching a new process, which is a bit hacky, this simply starts a new Qt app instance after the existing one exits. As well as fixing a bug (where the new process hangs on exit), this also makes the process far easier to debug, since there is no new process to attach to.