fix: Remove mutex locking in MSWindowsWatchdog (solves access violation bug)
Use of either C++11 mutex or native Win32 mutex causes access violation. The mutex was added as it looked safe to do so but perhaps we can live without it.

This probably a symptom of a much bigger problem like stack or heap corruption, making this more of a workaround than a proper fix.